LayoutTests/inspector/debugger/break-on-exception-expected.txt

CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testFinally
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: finally
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testThrowingThruNativeCode
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testCatch
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: catch TestError
CONSOLE MESSAGE: DONE
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Checking pause locations when pausing on all exceptions.


== Running test suite: BreakOnAnyException
-- Running test case: BreakOnAnyException.UncaughtExceptionBasic
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnAnyException.UncaughtExceptionInFinally
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnAnyException.UncaughtExceptionThruNativeCode
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnAnyException.CaughtException
PASS: Should pause for exception.
PAUSE AT doThrow:8:22

-- Running test case: BreakOnAnyException.Options.Condition

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: BreakOnAnyException.Options.IgnoreCount

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: BreakOnAnyException.Options.Action.Log

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: BreakOnAnyException.Options.Actions.Evaluate

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
